refactor: decompose main.py into agent_components modules
- Split main.py (1613 lines) into focused modules under agent_components/
- Main.py reduced to 624 lines containing core agent logic
- Created 10 specialized modules for better separation of concerns
- Added backward compatibility exports for smoother migration
- Updated refactoring plan documentation with completion status

Note: Tests need updating to import from new module locations.
This is tracked in the refactoring plan for next developer.
